refactor: created actions file for about us page
